Final phasing out of acs_named_objects
 - remove table creation
 - update documentation
 - provide an optional upgrade script to delete the table and the object type (commented out by default, as people might have defined object subtypes downstream).

Context: this table and the related machinery were created at a time where acs_objects where not as mature as they are now and would not include the "title" information. To retrieve the object name, the only option was the stored procedure acs_object.name, which was not scalable.

From an excerpt of the Changelog file:

.....

2004-03-31 jeffd

......
        * packages/categories/tcl/category-trees-procs-oracle.xql:
        * packages/categories/tcl/category-trees-procs-postgresql.xql:
        acs_named_objects superflous now that title is in
        acs_objects.  need to fix the rest of the occurances though
......

This tells us that the phasing out has already been decided back in 2004, but delayed. One technical reason, could be that the acs_named_objects table would also materialize the translation, while the title column would store sometimes a message key, that would need to be translated manually. This is helpful when e.g. sorting by name in sql. However, only the faq and categories package would seem to make any use of the table and only the faq package would seem to maintain its values... This would make some UIs in the categories package practically useless since a long time.